requirements.txt
paho-mqtt>=1.6.0
adafruit-circuitpython-neopixel-spi
spidev
